Move to plugin-wide common package
