+++ plugins/common/auth/basic_auth.go
@@ -1,4 +1,4 @@
-package common
+package auth
 
 import (
 	"crypto/subtle"
