+++ plugins/common/auth/basic_auth_test.go
@@ -1,4 +1,4 @@
-package common
+package auth
 
 import (
 	"github.com/stretchr/testify/require"
